feat: Add Footer component Closes #34
This commit adds a new Footer component to the project. The Footer component displays the copyright year and the developer's name. It also includes a link to the developer's website. This component improves the overall user experience by providing important information and giving credit to the developer.

Note: This commit message follows the conventional format of "feat: <description>".